[2024년 02월 정기 업데이트]

+++ inc/targets/st/class_gpio_stm32.h
@@ -44,7 +44,7 @@ public:
 
 	void setPackageAsAltFunc(AltFunc *altport, uint8_t numOfPort, uint8_t ospeed, uint8_t otype);
 
-	error setAsAltFunc(uint8_t pin, uint8_t altFunc, uint8_t ospeed = define::gpio::ospeed::MID, uint8_t otype = define::gpio::otype::PUSH_PULL);
+	error setAsAltFunc(uint8_t pin, uint16_t altFunc, uint8_t ospeed = define::gpio::ospeed::MID, uint8_t otype = define::gpio::otype::PUSH_PULL);
 
 	void setAsInput(uint8_t pin, uint8_t pullUpDown = define::gpio::pupd::NONE);
